Qt: Combine EmuThread.cpp and QtHost.cpp
These mostly called back between one another.
