+++ pcsx2-qt/Settings/AdvancedSystemSettingsWidget.cpp
@@ -19,7 +19,7 @@
 #include <algorithm>
 
 #include "AdvancedSystemSettingsWidget.h"
-#include "EmuThread.h"
+#include "QtHost.h"
 #include "QtUtils.h"
 #include "SettingWidgetBinder.h"
 #include "SettingsDialog.h"
