+++ pcsx2-qt/Settings/AudioSettingsWidget.cpp
@@ -19,7 +19,7 @@
 #include <algorithm>
 
 #include "AudioSettingsWidget.h"
-#include "EmuThread.h"
+#include "QtHost.h"
 #include "QtUtils.h"
 #include "SettingWidgetBinder.h"
 #include "SettingsDialog.h"
