+++ pcsx2-qt/Settings/DEV9SettingsWidget.cpp
@@ -26,7 +26,7 @@
 #include "pcsx2/HostSettings.h"
 
 #include "DEV9SettingsWidget.h"
-#include "EmuThread.h"
+#include "QtHost.h"
 #include "QtUtils.h"
 #include "SettingWidgetBinder.h"
 #include "SettingsDialog.h"
