+++ pcsx2-qt/Settings/SystemSettingsWidget.cpp
@@ -20,7 +20,7 @@
 
 #include "pcsx2/HostSettings.h"
 
-#include "EmuThread.h"
+#include "QtHost.h"
 #include "QtUtils.h"
 #include "SettingWidgetBinder.h"
 #include "SettingsDialog.h"
